refactored barcode detection code to be more extensible.
