+++ Binner/Binner.Web/ClientApp/public/image/barcode/examples/Stackpole-SEI-RMCF2010ZTOR00.json
@@ -1,0 +1,15 @@
+{
+  "type": "code128",
+  "labelType": "part",
+  "vendor": "Generic",
+  "value": {
+    "partNumber": "RMCF2010ZT0R00,Q4000,1TT170419155,17/16,JP"
+  },
+  "correctedValue": "1PRMCF2010ZT0R00,Q4000,1TT170419155,17/16,JP\r",
+  "rawValue": "1PRMCF2010ZT0R00,Q4000,1TT170419155,17/16,JP\r",
+  "rsDetected": false,
+  "gsDetected": false,
+  "eotDetected": false,
+  "invalidBarcodeDetected": false,
+  "rawValueFormatted": "1PRMCF2010ZT0R00,Q4000,1TT170419155,17/16,JP␍"
+}
